Add statically linked libnuma support for AOT
