+++ src/coreclr/gc/unix/CMakeLists.txt
@@ -5,6 +5,7 @@ include(configure.cmake)
 
 set(GC_PAL_SOURCES
     gcenv.unix.cpp
+    numasupport.dynamic.cpp
     events.cpp
     cgroup.cpp)
