+++ src/coreclr/nativeaot/Runtime/CMakeLists.txt
@@ -263,3 +263,8 @@ if(NOT CLR_CMAKE_TARGET_ARCH_WASM)
 else()
   add_subdirectory(Portable)
 endif()
+
+add_library(numasupportdynamic STATIC ${GC_DIR}/unix/numasupport.dynamic.cpp)
+install_static_library(numasupportdynamic aotsdk nativeaot)
+add_library(numasupportstatic STATIC ${GC_DIR}/unix/numasupport.static.cpp)
+install_static_library(numasupportstatic aotsdk nativeaot)
